requirements.txt
aiogram==3.13.0
aiohttp==3.10.5
python-dotenv==1.0.1
Pillow==10.4.0
rembg==2.0.59
onnxruntime==1.19.2
xai_sdk==1.11.0
pydantic>=2.0
